// cordobacelulares/AngularBoostrapCordobaCelulares/cordobacelulares/src/app/models/phone.ts
export interface Phone {
  readonly idPhone: number;

  version: string;
  images: string[];
  red: string,
  mainCamera: string[];
  secondaryCamera: string;
  screen: string[];
  processor: string;
  gpu: string;
  memory: string[];
  expansion: string;
  os: string;
  battery: string[];
  connectivity: string[];
  dimensions: string;
  security: string[];
  colors: string[];
  boxContents: string[];
  videoYoutube:string;
  brand: string;
  model: string;
}